functions/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
firebase-admin==6.4.0
firebase-functions==0.1.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
pytz==2024.1
mangum==0.17.0
pytest==7.4.4
pytest-asyncio==0.23.3